+++ scripts/quality-baseline.json
@@ -7,6 +7,6 @@
   "viewRuntimeBridgeLookups": 0,
   "labStateAppFiles": 0,
   "labStateTestFiles": 0,
-  "largeJsFilesOver800Lines": 6,
+  "largeJsFilesOver800Lines": 5,
   "maxJsFileLines": 1168
 }
